Turn off fatal warnings
We need to turn them off so that we can get to a "cannot possibly match" error that
is emitted afterwards. The error is emitted when we turn on the

    import Lst.::

imports. `Lst.::` is a poison pill that causes an error when we try to match a Lst with a ::
Withour that trick errors go unreported and the patterns simply do not match at runtime.