Make tasty inspector receive tasty files directly
We add ways to load a list of tasty files and all tasty files in a jar.
This change is aligned with the changes in progress on the `-from-tasty` compilation.

We remove the old way to load tasty file by passing the class name.
Loading by name had some unintended consecuences that made this API fragile.
 - Loaded Scala 2 classes by mistake (no tasty)
 - Loaded java classes by mistake (no tasty)
 - Load the wrong version of the tasty file because the classpath contains an older/newer version of that class (different tasty)