Scala.js: Implement JS exports.
Compared to the implementation in Scala 2, there is one significant
difference. In Scala 2, top-level exports and static exports info
are computed by `PrepJSInterop` and stored in an out-of-band
mutable Map until the back-end. In this implementation,
`PrepJSInterop` only performs checks for compile error messages,
and the required data are recomputed by the back-end from the
annotations.

Since dotc does not (currently?) constant-fold expressions within
annotations, `PrepJSInterop` specifically constant-folds the
arguments of `@JSExportTopLevel` and `@JSExportStatic`, so that the
literal values can be recovered by the back-end.