Fix dokka code
There was something really funky going on. Basically, we had

    extension (sym: Symbol)
      def documentation = sym.documentation match ...

How is this not an infinite loop? Well, name resolution tried the current method,
failed because of a cyclic reference and then fell back to an alternative
extension method elsewhere. So compilation relied crucially on a missing result type
in `documentation`. This is terrible! This PR plugs the specific hole, but
I fear there are others that could be exploited in user code.

Morale: Leaving out result types has an upside in that it detects compiler
bugs. But for robust software, it's better do add return types to all public methods.