Make the dynamic methods of reflect.Selectable final.
By making them final, the Scala.js back-end can process them not
only for `reflectiveSelectable`, but for any custom subclass of
`reflect.Selectable`. This removes its main limitation on Scala.js.

We add `@inline` to the default implementation in
`reflectiveSelectable` (which requires to make it a named class) so
that the Scala.js optimizer can completely optimize it away at
link time. On the JVM, this change has no effect.