Make parameterless givens lazy vals
We used to represent them as defs and converted some of them to lazy
vals later, in CacheAliasImplicits.

But that means a parameterless given value is never a stable prefix, so
we cannot use it as a qualifier for an abstract type.

We now change the scheme so that parameterless givens start out as lazy vals
and some of them are changed to defs later in CacheAliasImplicits.

This is not a uniform improvement, since lazy vals have the restriction
that their types must be realizable. See i9103.scala, where we had to
change a type field to a trait to keep it working.

But I think overall this is an improvement to what we had before.