community-build: Hardcode the version of some projects
Some projects rely on plugins (sbt-dynver, sbt-git) to compute a version
number that includes the current git commit sha, this makes the
community build really fragile since we need to hardcode the version
numbers of published projects, and that number will change everytime we
add a commit to a project. Moreover, `git submodule update` doesn't
fetch new tags and sbt-dynver relies on the latest to set the version
number, so we can end up in confusing situations where we're not
publishing the version we think we're publishing. Hardcoding version
numbers avoids this problem.

An alternative solution would be to get the version number from the sbt
project, but I don't know any good way to do that, parsing the output of
`sbt "show version"` to get just the version number doesn't seem easy. I
guess we could inject a new task that writes the version number to a
file on disk.

Revert the previous commit since we no longer need to fetch tags and
so that would just slowdown the CI for no benefits.