Fix #8001: Erase polymorphic value classes like Scala 2
Previously given:

    class Poly[A](val value: A) extends AnyVal

We always erased `Poly[X]` to `Object`, no matter the value `X`, because
the erasure was the erased underlying type as seen from its definition,
and not as seen from the current prefix. But it turns out that in Scala
2, `Foo[Int]` will be erased to `Integer` instead (it would have made
more sense to use `int` but I suspect this is more accidental than
designed).

To be binary-compatible with Scala 2 and to support the same kind of
overloads we need to replicate its behavior, more precisely the rules I
was able to reverse-engineer are:

- Given `class Foo[A](x: A) extends AnyVal`, `Foo[X]` should erase like
  `X`, except if its a primitive in which case it erases to the boxed
  version of this primitive.
- Given `class Bar[A](x: Array[A]) extends AnyVal`, `Bar[X]` will be
  erased like `Array[A]` as seen from its definition site, no matter
  the `X` (same if `A` is bounded).

I was able to adapt our implementation of value class erasure to these
new rules without too much refactoring through one compromise: a value
class can no longer wrap another value class. This was never supported
by Scala 2 so we can afford to not support it either.