Remove the -Yemit-tasty-in-class option, along with handling of .hasTasty.
The default was flipped in a6455665fa30b760051238df52c05de5e6ccd5b1
in July 2018. Since then, it has not been really tested.

Completely removing it will allow the tooling not to worry about
`.hasTasty` and the need to read TASTy from class files.

We keep the support of `.hasTasty` files in sbt-dotty for now, so
that it still supports older versions of dotty.