Remove reflect Constant.value
To use this we usually need to use an `asInstanceOf` which cannot be checked in the case of ClassOf. Doing so without being careful will introduce an unsoundness. It seems better to remove this method and let users match all the cases.

In the future, we could reintroduce a restricted version of this API that only works on a subset of the constants.