Remove MissingCoreLibTests
This test printed an error message in the log output which was extremely
misleading since it looked like a weird test failure. Removing it
instead of fixing it because it's not actually testing anything useful:
it's supposed to check that we don't get a cascade of error messages if
scala3-library is missing from the classpath, but after
24eb6cf13d603ff41511f85bb20147c5223bb656 this only worked by chance
because the code we're compiling is extremely simple (Foo.scala only
contains "class Foo"), so it's doubly misleading.