Stop using scala.runtime.function.JFunction*mc*
Instead, use the equivalent classes in the Scala 2 standard library in
scala.runtime.java8.

The only difference is that these classes are written in Scala instead
of Java, which exposed a bug in the Scala2Unpickler: we were calling
`atPhaseNoLater(picklerPhase)`, but this doesn't actually do anything in
situation where the Pickler phase is not used (for example, when using
the `QuoteCompiler`), this is fixed by introducing a new
`atPhaseBeforeTransforms` which does the right thing.

Because the `FirstTransform` phase itself might not be present, this
also required fixing `Phase#next` and `Phase#prev` to work on `NoPhase`:
the issue was that `myBase.NoPhase` failed with an NPE which we fix by
making `NoPhase` a static object instead of a member of `PhasesBase`.
Since an instace of `Phase` contains mutable state this requires making
it `@sharable` to pass our tests, this is safe since we never mutate
anything in `NoPhase` (this also matches what we do for `NoSymbol` and
`NoDenotation`).