fix #10857: export from wildcard more definitions
a wildcard export will first search for all public and
non-synthetic members to add. Then once a valid member
is found, that name is then used to search for all
term or type members that use that name and are not
given or private.

As a consequence members like companion objects to
case classes will be exported alongside a type alias
to the case class