Fix handling of default arguments in parameter-dependent using clauses
We had a problem in handling using clauses that were both parameter dependent
and had default parameters. Example:

    trait A:
      type X

    f(using a: A = defaultA, b: a.X)

If implicit search fails for `A`, we should search the second parameter with
`defaultA.X` as type. We searched instead with an error type (i.e. the type of
the missing argument for A). Sine error types are compatible with everything,
any implicit candidate would match. In the original test i5427 we had just one
closest nested candidate that was the right one, but once we add another
candidate with a different type next to the first one the test fails.