Add overriding check
In a situation like
```
class A { type T = B }
class B extends A { override type T }
```

We only checked that `T` in `A` correctly overrides `T` in `B`, since `T` in `A`
is taken to be the overriding member, since it is abstract. But we also have to check
the reverse.

This replaces a more specialized treatment where we checked that final members
in `Object` could not be overridden by abstract members in traits. I removed a bunch
of tests that all tested variations of that other check, and changed the check
file for one representative.

Fixes #11130