syntax.md: support multiple if-guards on same line
Consider the example:
```scala
for {
  x <- 1 to 5
  if true; if true; if true
} yield x
```
Note, that multiple if guards start from a new line and are not separated with a semicolon or a new line.
This code is compiled ok in Scala 3 and is a valid code according to Scala 2 syntax reference:
(though looks like there is a bug in implementation https://github.com/scala/bug/issues/12323)

But it's not supported according to the Scala 3 syntax reference.

related issues:
https://github.com/scala/bug/issues/12323
https://youtrack.jetbrains.com/issue/SCL-13220