Make it possible to always set the prefix of denotations
The comment above `derived` used to say "Currently this fails
bootstrap", this was fixed by #11303, and this commit takes care of
updating the few places where we were assuming that something had to be
a SymDenotation when it didn't. Nevertheless, we still keep essentially
the same behavior of reusing denotations that existed before this commit
for performance reason, but this is now toggleable using
`Config.reuseSymDenotations`.

There's one difference with what we did before: if we need to create a
new denotation anyway because the info changed, then we always update
the prefix at the same time, this doesn't cost us anything and means
that we can rely on `prefix` being correctly set on non-sym denotations.