Set defTree properly
In the following test:

- tests/pos/i3130b.scala

If we add `transparent`, then everything is OK.
The reason is that we set `Symbol.defTree` systematically
in PostTyper. Now the inlining happens after PostTyper,
thus `defTree` is not properly set for inlined definitions.

To compensate, we ensure that `defTree` is set in the ReTyper.
The IninerTyper extends ReTyper, thus it fixes the problem.
Doing the fix directly in InlinerTyper, however, does not pass
the CI. The reason is that `-Ycheck:all` will run `TreeChecker`
which will make tree bindings get lost.