Fix tests/pos/i5655
The commits 5cf9e0b incorrectly throw the prefix away
and construct the type from the symbol by ignoring
type parameters of generic classes.

As prefix is already in the "inner normal form",
it suffices to select member on it. The commit 22ab968
calls `processInner`, which is effectively an no-op.

Also found that we should reset `skiptvs` in reading
method parameter and result types.