Refactor class parents handling
Before this commit we had:
- Type#parents: for a ClassInfo, this is the parents of the class as seen from
  the ClassInfo prefix
- ClassInfo#classParents: this is also the parents of the class but
  _not_ as seen from the ClassInfo prefix
- ClassDenotation#classParents: this delegates to ClassInfo#parents and
  not to ClassInfo#classParents as one might think.

To clear things up this commit:
- Renames ClassInfo#classParents to declaredParents, to avoid accidentally
  using it instead of parents.
- Remove ClassDenotation#classParents, one can always do `info.parents`
  instead.

Additionally, this commit introduces ClassDenotation#parentSyms, to be
used when one only care about the symbols of the parents. This method
relies on `declaredParents` instead of `parents` which avoids any
unnecessary call to `asSeenFrom`. It is used in `computeBaseData`
instead of the old `classParents` to avoid a cyclic reference issue I
ran into while working on an unrelated PR.