SingleDenotation#signature: Use Scala 3 signature for non-SymDenotations
For SymDenotations, it's important that we use a signature that matches
how the type is erased to maintain the invariant that two overloads in
the same owner have different signatures, but for non-SymDenotations we
don't make this guarantee so we can choose whatever we want as long as
it's consistent.

I originally chose to treat them like SymDenotation for consistency, but
we're about to add some complex (and therefore somewhat expensive) logic
for dealing with Scala 2 intersection erasure, by using the Scala 3
signature for all non-SymDenotation we avoid calling this logic more
than we need to.