Improve skipping on syntax errors
Simplify and improve the algorithm used for skipping tokens when a syntax
error is encountered. The new algorithm is based on the region stack maintained
in Scanner instead of maintaining separate counters for open parentheses. This
improves the precision of skipping in general, since there's is no chance that
the two datastructures get out of sync. It seems overall we produce fewer follow
on error after a syntax error that way.