Support Scala.js unions extra methods via implicit conversion
The companion of `js.|` contains the following widely used implicit
conversion to a value class:

    /** Provides an [[Option]]-like API to [[js.UndefOr]]. */
    implicit def undefOr2ops[A](value: js.UndefOr[A]): js.UndefOrOps[A] =
      new js.UndefOrOps(value)

(where `UndefOr[A]` dealiases to `A | Unit`).

Since we re-interpret Scala.js unions as real unions, this companion is
not in the implicit scope of `A | Unit`, we work around this by
injecting a new `UnitOps` with the implicit conversion we want in the
implicit scope of `Unit` (we could have directly injected the object
`js.|`, but that contains other implicits we do not need).

This finally lets us compile and run the sjsJUnitTests after the
previous two commits.