Insert traits with implicit parameters as extra parents of classes
If a trait A extends some other trait B which takes an implicit parameter, and
a class C extends A but not B, we need to insert B into the parents of A, so
that the implicit argument can be generated.

This required a generalization of the syncing between parent trees and parent types.
Namer will augment the parent types of a class with possibly other types.
Previously the only such change was a possibly leading class type, but we now
also generate trait parents that take context parameters. The new method parentTrees
reflects any such changes in the parent trees in Typer.