Scaladoc: add Wiki code blocks to Md syntax
The idea here is that in the wild, the parts of Wiki syntax that are the most
used are lists, code blocks, monospace text, and links. With this PR we now
more-or-less support all of them in Markdown syntax, which means that one will
be able to use it for old doc comments and get acceptable-looking output.