Better handling of leading infix operators in indented code ```code @main def Test =   val x = false   val y = 1   val result =     x     || y.match         case 1 => false         case 3 => false         case _ => true     || !x   assert(result) ``` In this code, the last `|| !x` was seen as a part of the previous case, so the code was parsed as ```scala @main def Test =   val x = false   val y = 1   val result =     x     || y.match         case 1 => false         case 3 => false         case _ => true || !x   assert(result) ``` This is highly surprising and unintuitive. The fix will insert an <outdent> token instead if the leading infix operator is too far to the left. Too far means: (1) left of the current indentation region, (2) and not to the right of any outer indentation widths.
(2) allows to still parse code like this
```scala
if xyz then
    one
  + two
  + three
```
Here, the width of the indentation region after `then` is 4, but the `+` operator is to the right
of the outer indentation width of 0., so the indentation region is not closed. In other words,
we do not close an indentation region if the result would not be legal, since it matches none of
the previous indentation widths.