Reset GADT constraint on failing isSubType
The normal constraint gets reset when isSubType fails - should the GADT constraint
not get reset as well? Subtyping is an exploration with backtracking where
some branches do not lead to a result. It seems unsound that failing explorations
can nevertheless introduce new GADT bounds.

I tried to change this in this commit, but it makes the following tests
fail with stack overflows:

    tests/pos/i9740b.scala failed
    tests/pos/i9740c.scala failed

I must admit I am not quite sure about whether it is enough to reset just the constraint
part of a GadtConstraint, or whether the whole constraint has to be copied. I hope not the
latter, since that would impose a heavy complexity price on subtype checking.