Refine resultTypeApprox
Add a special case to `resultTypeApprox` for method types where
the parameters appear only co-variantly in the result.

This allows types to be propagated enough to make `summon` work as well as
`implicitly`. Note that `summon` is result-dependent, but `implicitly` is not:

    def summon[T](using x: T): x.type = x
    def implicitly[T](implicit x: T): T = x

The tweak to `resultTypeApprox` makes `summon` behave like `implicitly`
for the purposes of type inference.

Fixes #12005