Fix #11885: Always clone trees if position already set
The cause of the problem is that we use parallelism in Pickler:

      if !Pickler.ParallelPickling || ctx.settings.YtestPickler.value then force()

Sometimes on Windows, the futures run a little slower, and the later
phase `Inlining` can change the positions of the trees to be pickled,
thus non-determinism.

For the Dotty project, the statistics is as follows:

- Before: ntrees = 5331539
- After:  ntrees = 5334075

Performance-wise, this should be better than synchronizing the pickling tasks.