Fix signatures involving WildcardTypes
A `WildcardType` never appears in the type of a tree, but it can appear
in an expected type and because TypeComparer checks if
`tp1.signature consistentParams tp2.signature`, we can end up calling
`sigName` on a `WildcardType`.

Before this commit, the result was either a custom type name or the
upper-bound of the wildcard, but both of these options means that
`consistentParams` could return false in situations where the two method
types would in fact match. We fix this by always returning
`tpnme.Uninstantiated`, meaning that `consistentParams` will always
allow a wildcard to match any other type. This does not cause any
over-approximation because the TypeComparer will always check that the
actual types match after checking that the signatures match.

Also use tpnme.ERROR instead of tpnme.Wildcard for ErrorType and NoType
to make it easier to spot their usage.

Fixes #11481.