Peformance test for WIP
Use these changes to compile the stdlib. You need to comment out AnyVal.scala and then
use the following command on community-projects/stdLib213:

    sc -Ydetailed-stats *.scala */*.scala */*/*.scala */*/*/*.scala

What I see is:

Calls to `translucentSuperType` 53152:

    try norm KO -> 53086
    try norm OK -> 66

The only computationally relevant part of this are calls to tryMatchAlias

    try match alias -> 528

Those calls suceed in 66 cases

    try norm OK -> 66

The successful normalizations all return types like this:

    try norm OK F[T1] *: scala.Tuple.Map[(T2, T3, T4, T5, T6, T7, T8), F] -> 3

Now it could be that trying the normalizations is costly, or that the normalized types
cause exhaustivity checking to explode. Tht remains to be found out.