Fix renaming of constraints
When we replace a type lambda in a constraint by another,
the type variables associated with the old lambda changed their origin
to point to the new lambda. But that means that the same type
variables in predecessor typestates do not point to their associated
lambda anymore. To fix this, we have to do the same renaming
of type lambdas also in predecessor typestates.