Make scaladoc a bootstrapped-only project
Previously, we had both a non-bootstrapped and bootstrapped version of
that project as well as its dependencies (scala3-library-js,
scala3-tasty-inspector, scaladoc-testcases, scaladoc-js), the main
reason for this is that it allowed the non-boostrapped scaladoc to be
part of the `scalaInstance` of all bootstrapped projects, allowing us to
run `scala3-library-bootstrapped/doc` for example. But this also meant
that simply trying to run the compiler would compiler scaladoc as well
as its dependencies even if no documentation needed to be generated, and
since `scaladoc-js` runs the Scala.js optimizer this could be quite slow.

This commit fixes that by using a separate `scalaInstance` for the `doc`
task, which means the regular `scalaInstance` no longer needs to depend
on scaladoc and also allows us to remove all the non-bootstrapped
projects related to scaladoc (thus simplifying the build a lot) since we
can safely use the bootstrapped version in `doc / scalaInstance` without
running into a loop.