Disallow experimental CB projects on release builds
Previous mechanism for this inferred experimental capabilities
from the version of the compiler (SNAPSHOT or NIGHTLY permits
experimental). However this doesn't work in practice:
the version of the compiler is set by CI to a stable version only
when publish task is executed (not when test tasks are executed).
If we set the version as stable for community build as well,
the community projects will not be able to resolve it as they operate
against the locally published snapshot and not the official stable
release version.

Hence this solution adds another env var specifically to track
experimental features allowance.