sbt-dotty: Fix the binary suffix used for publishing
The compiler we publish is compiled with a non-bootstrapped compiler, so
its scalaBinaryVersion suffix is computed from the version number of the
non-bootstrapped compiler, for 3.0.0-RC3 this was
`3.0.0-RC3-nonboostrapped` and we correctly return `3.0.0-RC3` as a
binary version, but for 3.0.0 this is `3.0.0-nonbootstrapped` for which
we return again `3.0.0-nonbootstrapped` when we really want just `3`.
This is why we incorrectly published https://repo1.maven.org/maven2/org/scala-lang/scala3-compiler_3.0.0-nonbootstrapped/

Fixed by special-casing RCs and milestones which are the only special
suffixes that get their own binary suffix.

This fix is specific to the release-3.0.0 branch since the master branch
does not use sbt-dotty anymore (and will need its own fix to publish
3.0.1-RC1 correctly due to a different issue with how sbt computes
the suffix for `3.0.1-RC1-nonbootstrapped`).