Refactor domain: Introduce Warm and ThisRef
This brings us closer to OOPSLA paper.
The key insight to avoid cycles like the following

```
class A(b: B) {
  val b2 = new B(this)
}

class B(a: A) {
  val a2 = new A(this)
}
```

is to use `Warm(C, outer)` to evaluate fields and methods of `C`.
The evaluation results are cached uniformly. Therefore, there is
no need to treat `Warm(C, outer)` as addresses.