Don't instantiate variables to Nothing in canDefineFurther
The purpose of canDefineFurther is to discover new members. That is not
helped by instantiating a type variable to `Nothing`. Avoiding the instantiation
helpos error messages. There was one test (i864.scala) that fails now
but compiled previously with a surprising instantiation to Nothing. I believe
it is better to issue an error message in this case.